libcod: ratelimit SV_DirectConnect()
It's still possible to flood this (tested via script)
Also make Sys_Milliseconds64() a separate function

+++ declarations.hpp
@@ -1,8 +1,6 @@
 #ifndef _DECLARATIONS_HPP_
 #define _DECLARATIONS_HPP_
 
-#include <stdint.h>
-
 #define qboolean int
 #define qtrue 1
 #define qfalse 0
@@ -660,7 +658,7 @@ struct leakyBucket_s
 {
 	netadrtype_t type;
 	unsigned char adr[4];
-	uint64_t lastTime;
+	int lastTime;
 	signed char	burst;
 	long hash;
 	leakyBucket_t *prev, *next;
